pyproject.toml
[project]
name = "finance"
version = "0.1.0"
description = "Personal finance tracker with SimpleFIN sync and MCP server"
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "httpx",
    "click",
    "python-dotenv",
    "fastapi",
    "jinja2",
    "uvicorn",
    "python-multipart",
    "anthropic",
]

[project.scripts]
finance = "finance.cli:main"
finance-mcp = "finance.server:main"
finance-dashboard = "finance.web.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
